requirements.txt
langchain-core
langgraph
langchain-openai

streamlit
fastapi

langchain-community

python-dotenv
pydantic
requests
